chore: update @planet-sdk/common to version 0.2.4 and refactor imports
- Updated package-lock.json and package.json to use @planet-sdk/common version 0.2.4.
- Refactored imports across multiple files to source types from @planet-sdk/common instead of local paths.
- Removed obsolete intervention.d.ts file as its types are now sourced from @planet-sdk/common.
- Adjusted related components and contexts to ensure compatibility with the updated type definitions.
