+++ src/features/projectsV2/ProjectDetails/components/SingleTreeInfo.tsx
@@ -1,8 +1,8 @@
 import type { SetState } from '../../../common/types/common';
 import type {
-  SingleTreeRegistration,
   SampleTreeRegistration,
-} from '../../../common/types/intervention';
+  SingleTreeRegistration,
+} from '@planet-sdk/common';
 
 import PlantInfoCard from './microComponents/PlantInfoCard';
 import styles from '../styles/InterventionInfo.module.scss';
