+++ src/features/user/TreeMapper/Import/components/SampleTreeCard.tsx
@@ -1,7 +1,7 @@
 import type { ReactElement } from 'react';
-import type { MultiTreeRegistration } from '../../../../common/types/intervention';
-import type { SampleTree } from '../../../../common/types/intervention';
+import type { SampleTree } from '../../Treemapper';
 import type { Control, FieldArrayWithId, FieldErrors } from 'react-hook-form';
+import type { MultiTreeRegistration } from '@planet-sdk/common';
 
 import { useTranslations } from 'next-intl';
 import styles from '../Import.module.scss';
