+++ src/features/user/TreeMapper/MySpecies/MySpeciesForm.tsx
@@ -1,8 +1,6 @@
 import type { APIError } from '@planet-sdk/common';
-import type {
-  Species,
-  SpeciesSuggestionType,
-} from '../../../common/types/intervention';
+import type { SpeciesSuggestionType } from '../Treemapper';
+import type { Species } from '../Treemapper';
 
 import { useEffect, useState, useContext } from 'react';
 import StyledForm from '../../../common/Layout/StyledForm';
