+++ src/features/user/TreeMapper/Treemapper.d.ts
@@ -19,3 +19,28 @@ export interface SpeciesFormData {
   otherSpecies: string;
   treeCount: number;
 }
+
+export interface Species {
+  aliases: string;
+  description: string | null;
+  id: string;
+  image: string | null;
+  scientificName: string;
+  scientificSpecies: string;
+}
+
+export interface SpeciesSuggestionType {
+  id: string;
+  name: string;
+  scientificName: string;
+}
+
+export type SampleTree = {
+  plantingDate: Date;
+  treeTag: string;
+  height: string;
+  diameter: string;
+  otherSpecies: string;
+  latitude: string;
+  longitude: string;
+};
